add suggested changes from  code review
