+++ lib/recipes.js
@@ -6,7 +6,7 @@ class RecipeClass {
     this.id = id;
     this.title = title;
     this.subtitle = subtitle;
-    this.picture = picture || "assets/recipe-placeholder.png";
+    this.picture = picture || "recipe-placeholder.png";
   }
 }
 
@@ -15,26 +15,21 @@ const recipes = [
     1,
     "Sojabällchen-Curry",
     "Rezept aus Indien, mit Reis und Kartoffeln",
-    "assets/recipe-1.png"
+    "recipe-1.png"
   ),
   new RecipeClass(
     2,
     "Schokoladenkuchen",
     "Vegan und ohne Mehl, einfach gemacht",
-    "assets/recipe-2.png"
+    "recipe-2.png"
   ),
   new RecipeClass(
     3,
     "Kürbis-Ingwer-Suppe",
     "Mit Kokosmilch und Kürbiskernen",
-    "assets/recipe-3.png"
+    "recipe-3.png"
   ),
-  new RecipeClass(
-    4,
-    "Blaubeer-Crumble",
-    "Mit Vanilleeis",
-    "assets/recipe-4.png"
-  ),
+  new RecipeClass(4, "Blaubeer-Crumble", "Mit Vanilleeis", "recipe-4.png"),
 ];
 
 export { RecipeClass, recipes };
