+++ pages/index.js
@@ -1,4 +1,4 @@
-import Recipes from "../components/Recipes";
+import Recipes from "../components/Recipes/Recipes";
 
 export default function RecipesPage() {
   return <Recipes />;
